Move common utility fns to utils.go
