+++ controllers/replicationdestination_controller.go
@@ -134,21 +134,6 @@ func (r *ReplicationDestinationReconciler) SetupWithManager(mgr ctrl.Manager) er
 		Complete(r)
 }
 
-// reconcileFunc is a function that partially reconciles an object. It returns a
-// bool indicating whether reconciling should continue and an error.
-type reconcileFunc func(logr.Logger) (bool, error)
-
-// reconcileBatch steps through a list of reconcile functions until one returns
-// false or an error.
-func reconcileBatch(l logr.Logger, reconcileFuncs ...reconcileFunc) (bool, error) {
-	for _, f := range reconcileFuncs {
-		if cont, err := f(l); !cont || err != nil {
-			return cont, err
-		}
-	}
-	return true, nil
-}
-
 type rsyncDestReconciler struct {
 	destinationVolumeHandler
 	service    *corev1.Service
