+++ controllers/utils.go
@@ -1,0 +1,46 @@
+/*
+Copyright 2020 The Scribe authors.
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU Affero General Public License as published
+by the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU Affero General Public License for more details.
+
+You should have received a copy of the GNU Affero General Public License
+along with this program.  If not, see <https://www.gnu.org/licenses/>.
+*/
+
+package controllers
+
+import (
+	"github.com/go-logr/logr"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
+)
+
+func nameFor(obj metav1.Object) types.NamespacedName {
+	return types.NamespacedName{
+		Name:      obj.GetName(),
+		Namespace: obj.GetNamespace(),
+	}
+}
+
+// reconcileFunc is a function that partially reconciles an object. It returns a
+// bool indicating whether reconciling should continue and an error.
+type reconcileFunc func(logr.Logger) (bool, error)
+
+// reconcileBatch steps through a list of reconcile functions until one returns
+// false or an error.
+func reconcileBatch(l logr.Logger, reconcileFuncs ...reconcileFunc) (bool, error) {
+	for _, f := range reconcileFuncs {
+		if cont, err := f(l); !cont || err != nil {
+			return cont, err
+		}
+	}
+	return true, nil
+}
